aider: Added a default icon for resources with undefined or unexpected types in `ResourcesList.tsx`.
# Aider chat conversation:

USER: some results does not have icons, why is that

ASSISTANT: In the `ResourcesList.tsx` file, the icons are conditionally rendered based on the `type` of the resource. If a resource has a type of `'url'`, a `LinkIcon` is displayed, and if the type is `'pdf'`, a `DocumentTextIcon` is displayed. If the resource has a different type or the type is not set, no icon will be displayed.

Here's the relevant code snippet from `ResourcesList.tsx`:

```tsx
{resource.type === 'url' ? <LinkIcon className="w-8 h-8 mr-2" /> : resource.type === 'pdf' ? <DocumentTextIcon className="w-8 h-8 mr-2" /> : null}
```

If you are seeing resources without icons, it's likely because their `type` is neither `'url'` nor `'pdf'`. To ensure that all resources have icons, you would need to add additional conditions to handle other types or provide a default icon for resources with undefined or unexpected types.

If you want to add a default icon for all other types, you could modify the code like this:

```tsx
src/client/research/ResourcesList.tsx
<<<<<<< SEARCH
{resource.type === 'url' ? <LinkIcon className="w-8 h-8 mr-2" /> : resource.type === 'pdf' ? <DocumentTextIcon className="w-8 h-8 mr-2" /> : null}
=======
{resource.type === 'url' ? <LinkIcon className="w-8 h-8 mr-2" /> : resource.type === 'pdf' ? <DocumentTextIcon className="w-8 h-8 mr-2" /> : <DefaultIcon className="w-8 h-8 mr-2" />}
>>>>>>> REPLACE
```

In the `REPLACE` section, `DefaultIcon` is a placeholder for whatever default icon component you want to use. You would need to import this default icon at the top of your file, similar to how `LinkIcon` and `DocumentTextIcon` are imported.

Please provide the default icon you would like to use, or if you have other types that should have specific icons, let me know what they are and what icons should be used for them.