aider: Standardize the size of icons in the `ResourceItem` component to be 5x5.
# Aider chat conversation:

USER: make the icons in the same size

ASSISTANT: To ensure that all icons in the `ResourceItem.tsx` component are of the same size, we need to set a consistent width and height for each icon. Let's standardize the size of the icons to `w-5 h-5` (width 5, height 5) which is a common size for icons in Tailwind CSS.

Here's the change that needs to be made in the `ResourceItem.tsx` file:

```tsx
src/client/research/ResourceItem.tsx
<<<<<<< SEARCH
        <IconComponent className="w-8 h-8 mr-2" />
=======
        <IconComponent className="w-5 h-5 mr-2" />
>>>>>>> REPLACE

<<<<<<< SEARCH
        <TrashIcon className="w-7 h-7 mr-2" />
=======
        <TrashIcon className="w-5 h-5 mr-2" />
>>>>>>> REPLACE
```

This change will update the size of the icons used in the `ResourceItem` component to be 5x5, which should make them all the same size. If there are other icons in the file that also need to be resized, you can apply the same `className` change to them as well.